requirements.txt
django
djangorestframework
django-cors-headers
django-debug-toolbar
django-environ
dj-database-url
psycopg2
sendgrid
